Site working but tests failing
The goal of this eventual PR is to have the basic functionality of the
application to be fully tested before making modifications. I did some
basic restructuring and cleaning of style. I stubbed out the beginnings
of a test framework using py.test, but I can't seem to get py.test (or
anything for that matter) working.